\name{updateObject}
\alias{updateObject}
\alias{updateObject,GInteractions-method}
\title{Update a GInteractions object}
\description{Update an old GInteractions object to the latest format.}
\usage{
\S4method{updateObject}{GInteractions}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A old \link{GInteractions} object.}

\item{...}{Additional arguments that are ignored.}

\item{verbose}{Logical scalar indicating whether a message should be emitted as the object is updated.}
}
\value{
An updated version of \code{object}.
}

